refactor: split general controls
